cleared LsApp moved items and BalanceResult to items package
