+++ app/src/main/java/com/ivanchug/moneytracker/MoneyTrackerDbHelper.java
@@ -6,7 +6,7 @@ import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import com.ivanchug.moneytracker.api.Item;
+import com.ivanchug.moneytracker.items.Item;
 
 import java.util.ArrayList;
 import java.util.Date;
